__init__.py: try to do the right thing with the integrator
Jason recommends the LSODA integrator but I had had trouble getting it to work
in the past. I have dug into the details and this is my best effort at getting
it to work well. It turns out that it ignores tau, except for the tau_max
clipping; it also clips off zero intensity points at the edges of the
integration domain. Othert than that it seems to proceed in the way that you
would expect, except for the bit where it wants the X array to be backwards.
The other integrators don't have all of these bits, and some of them actually
use tau, I think, so I've split off a more generic function.